Completed sequencer
finished creating the sequencer class, it is currently very minimal and will need more features and improvements in the future

+++ Digital_sound_generation_project/Digital_sound_generation_project/Note.h
@@ -7,7 +7,7 @@ struct Note {
 	}
 	int frequency;
 	int startTime;
-	int volume;
+	float volume;
 	int generationType; //0: sin, 1: saw, 2: triangle, 3: square, 4: noise
 	int sampleIndex; //leave as -1 for none
 	Envelope envelope;
